+++ app.js
@@ -9,13 +9,7 @@ const errorsHandler = require('./middlewares/errors-handler');
 const routes = require('./routes');
 const { PORT, DB_URL } = require('./config');
 const { requestLogger, errorLogger } = require('./middlewares/logger');
-
-const limiter = rateLimit({
-  windowMs: 15 * 60 * 1000, // 15 minutes
-  max: 100, // Limit each IP to 100 requests per `window` (here, per 15 minutes)
-  standardHeaders: true, // Return rate limit info in the `RateLimit-*` headers
-  legacyHeaders: false, // Disable the `X-RateLimit-*` headers
-});
+const { limiter } = require('./middlewares/limiter');
 
 const app = express();
 // app.use(express.json());
